pokedex: implemented strongest pokemon table
First version of table with the strongest pokemon of each type. The page is not completed. There are some extra features already planned to be added. However, this is an already usable version.

Also, got rid of old code which assumed that this pokedex would include data about the regular games and not just pogo. The project has focused on pogo and I no longer plan on include non-pogo data (for now).