pokedex: window back and forward supported and IVs can be specified
This is a big update.

The dex is now using window states and updating the URL with search params to specify what is being loaded. This allows the browser's back and forward feature to work as expected. It is a huge improvement!

Also, added inputs to specify a pokemon's IVs and be able to calculate its information based on those specific stats.