pokedex: added new PvE metrics
There are three available metrics now and they can be changed from a new settings panel.

The Legend and Sources were updated accordingly and a new Credits section was added.

Also, a new footer with some contact info was added and the home link to my personal page was removed due to it being a bit confusing.